MWE to post about Eq for trait objects
